% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metdata.R
\docType{data}
\name{allTables}
\alias{allTables}
\title{Tables for the TUNA-VERSE package}
\format{
A character vector with 12 elements - 
MON_DOCS, TRIPS, PRO_SPC_INFO, LOG_EFRT_STD_INFO, LOG_SPC_STD_INFO, 
SLIP_OFFLD_STD_INFO, SLIP_BUYR_STD_INFO, SLIP_SPC_STD_INFO, LOG_EFRT_ENTRD_DETS, 
MON_DOC_ENTRD_DETS, LOG_SPC_ENTRD_DETS, AREAS
}
\description{
A character vector of table names used in the TUNAVERSE package.
}
