% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoFilter.R
\name{autoFilter}
\alias{autoFilter}
\title{autoFilter}
\usage{
autoFilter(tblList)
}
\arguments{
\item{tblList}{this is a vector of object names necessary for the functions of TUNAVERSE to run.}
}
\description{
This function operates on the various object identified in \code{tblList}, and 
ensures that they don't contain extraneous records.  For example, if the TRIPS object within 
tblList is filtered to only 2013 records, other objects such as MON_DOCS and LOG_SPC_STD_INFO 
will be filtered to ensure that only contain records that related to the remianing trips.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
