% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFT_COGNOS.R
\name{BFT_COGNOS}
\alias{BFT_COGNOS}
\title{BFT_COGNOS}
\usage{
BFT_COGNOS(QAQC = TRUE)
}
\arguments{
\item{QAQC}{Logical. If true (the default), the COGNOS report is QAQC'd. If false,
you get the original data with all its warts.}
}
\description{
This function loads the most recent extracted COGNOS data and
optionally returns a QAQC'd data object. At the moment there is no slick way to link and
extract reports from this southern Gulf of St. Lawrence fishery database.
}
\author{
Alex Hanke, \email{Alex.Hanke@dfo-mpo.gc.ca}
}
