% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFT_LOGS.R
\name{BFT_LOGS}
\alias{BFT_LOGS}
\title{BFT_LOGS}
\usage{
BFT_LOGS(tblList)
}
\arguments{
\item{tblList}{this is a vector of object names necessary for the functions of TUNAVERSE to run.}
}
\description{
This function operates on the various object identified in \code{tblList}, and 
reformats them to imitate the MARFISSCI.BFT_LOGS view, but valid for all MON_DOC types.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
