% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFT_LOGS_legacy.R
\name{BFT_LOGS_legacy}
\alias{BFT_LOGS_legacy}
\title{BFT_LOGS_legacy}
\usage{
BFT_LOGS_legacy(dsn = "PTRAN 64bit", uid = NULL, pwd = NULL)
}
\arguments{
\item{dsn, }{this is a character vector with the DSN name configured on the 
local computer allowing communication with the MARFIS tables.}
}
\description{
This function recreates a legacy MARFIS view that was used by the
Large Pelagics Group at SABs until 2023 when new logbooks affected the completeness
of the data extracted. It is recreated here for comparability purposes so that 
the data used in analyses prior to 2023 can be compared with data now proided by
the BFT_LOGS R function.
}
\author{
Alex Hanke, \email{Alex.Hanke@dfo-mpo.gc.ca}
}
