% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dropFields}
\alias{dropFields}
\title{dropFields}
\usage{
dropFields(tblList)
}
\arguments{
\item{tblList}{this is a vector of object names necessary for the functions of TUNAVERSE to run.}
}
\description{
By default, the original MARFIS objects have a number of fields that are not 
useful for the purposes of this package. This function removes them.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
