% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enWidener}
\alias{enWidener}
\title{enWidener}
\usage{
enWidener(tblList)
}
\arguments{
\item{tblList}{this is a vector of object names necessary for the functions of TUNAVERSE to run.}
}
\description{
A number of data objects used by this package store their data in "long" format 
(i.e. various *_DETS).  This function converts these objects from long to wide while only 
retaining fields directly related to the large pelagics fishery.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
