% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extraction_Date.R
\name{ExtractDate}
\alias{ExtractDate}
\title{ExtractDate}
\usage{
ExtractDate()
}
\description{
This function provides the date the Oracle tables were last extracted
and the number of elapsed days.
}
\author{
Alex Hanke, \email{Alex.Hanke@dfo-mpo.gc.ca}
}
