% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gearByYear.R
\name{gearByYear}
\alias{gearByYear}
\title{gearByYear}
\usage{
gearByYear(df)
}
\arguments{
\item{df}{this is a dataframe likely generated by BFT_LOGS().}
}
\description{
This function generates a tibble that lists how many occurrences of a variety of 
gear types were observed each year
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
