% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadIntoList}
\alias{loadIntoList}
\title{loadIntoList}
\usage{
loadIntoList(tblList, removeOriginals = F)
}
\arguments{
\item{tblList}{this is a vector of object names necessary for the functions of TUNAVERSE to run.}

\item{removeOriginals}{default is \code{F}.  Once the specified tables have been loaded into a 
list, this tells the function whether to retain or delete the original objects.}
}
\description{
This function takes a (specified) list of object names, and pulls them out of the 
global environment and into a list object.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
