% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP_LICENCES.R
\name{LP_LICENCES}
\alias{LP_LICENCES}
\title{LP_LICENCES}
\usage{
LP_LICENCES(
  type = "PLL_LICENCES",
  data_path = "R:/ATLSiteShares/SABS/LargePelagics/COMMERCIAL DATA/TUNAVERSE/data/"
)
}
\arguments{
\item{type}{default is \code{"PLL_LICENCES"}.  Valid values include "PLL_LICENCES", "MAR_BFT_LICENCES", "BFT_TRAP_LICENCES", "HARP_LICENCES".
this is a one element vector identifying the licence type which is matched against a list.}

\item{data_path}{default is \code{"R:/ATLSiteShares/SABS/LargePelagics/COMMERCIAL DATA/TUNAVERSE/data/"}. 
This is the location of the PLL_LICENCES, MAR_BFT_LICENCES, BFT_TRAP_LICENCES and HARP_LICENCES Rdata files.}
}
\description{
This function loads the specified licence data into the global environment.
Licence type is a proxy for fleet.
}
\author{
Alex Hanke, \email{Alex.Hanke@dfo-mpo.gc.ca}
}
