% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFT_LOGS_legacy.r
\name{BFT_LOGS_legacy}
\alias{BFT_LOGS_legacy}
\title{BFT_LOGS_legacy}
\usage{
BFT_LOGS_legacy(cxn = NULL)
}
\arguments{
\item{cxn}{A valid Oracle connection object. This parameter allows you to 
pass an existing connection, reducing the need to establish a new connection 
within the function. If provided, it takes precedence over the connection-
related parameters.}
}
\description{
This function recreates a legacy MARFIS view that was used by the
Large Pelagics Group at SABs until 2023 when new logbooks affected the completeness
of the data extracted. It is recreated here for comparability purposes so that 
the data used in analyses prior to 2023 can be compared with data now proided by
the BFT_LOGS R function.
}
\author{
Alex Hanke, \email{Alex.Hanke@dfo-mpo.gc.ca}
}
