% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LARGE_PELAGICS_LOGS_legacy.R
\name{LARGE_PELAGICS_LOGS_legacy}
\alias{LARGE_PELAGICS_LOGS_legacy}
\title{LARGE_PELAGICS_LOGS_legacy}
\usage{
LARGE_PELAGICS_LOGS_legacy(dsn = "PTRAN 64bit", uid = NULL, pwd = NULL)
}
\arguments{
\item{dsn, }{this is a character vector with the DSN name configured on the 
local computer allowing communication with the MARFIS tables.}
}
\description{
This function recreates a legacy MARFIS view that is used by the
Large Pelagics Group at SABs. This version is more robust in that it is able to 
address incorrect data format issues for some fields. It is the responsibility
of the analyst to find the inconsistent formats and correct them.
}
\author{
Alex Hanke, \email{Alex.Hanke@dfo-mpo.gc.ca}
}
