% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LP_LICENCES.R
\name{LP_LICENCES}
\alias{LP_LICENCES}
\title{LP_LICENCES}
\usage{
LP_LICENCES(
  type = "PLL_LICENCES",
  data_path = "R:/ATLSiteShares/SABS/LargePelagics/COMMERCIAL DATA/TUNAVERSE/data/"
)
}
\arguments{
\item{type}{this is a one element vector identifying the licence type which is matched against a list.}

\item{data_path}{this is the location of the PLL_LICENCES, MAR_BFT_LICENCES and HP_LICENCES Rdata files.}
}
\description{
This function loads the specified licence data into the global environment.
}
\author{
Alex Hanke, \email{Alex.Hanke@dfo-mpo.gc.ca}
}
